实现RAS加密和解密校验，并通过sa-token保存登录信息

+++ src/main/java/com/pn/news/model/pojo/User.java
@@ -125,7 +125,7 @@ public class User {
      * 取值constant中的平台
      */
     @TableField(exist = false)
-    private byte platform;
+    private int platform;
 
     /**
      * 验证码
